Moved logging setup to `commanderbot.core`
